+++ server/src/controllers/user.ts
@@ -1,9 +1,13 @@
 import createHttpError from "http-errors";
 import asyncWrapper from "../helpers/asyncWrapper";
+import * as userPermission from "../permissions/user";
 import * as userService from "../services/user";
 
-export const getUsers = asyncWrapper(async (_req, res) => {
+export const getUsers = asyncWrapper(async (req, res) => {
   const users = await userService.getAllUsers();
+
+  if (!userPermission.canGetAll(req.user))
+    throw createHttpError(401, "You don't have access to this resource");
 
   res.status(200).json({ users, count: users.length });
 });
@@ -11,15 +15,20 @@ export const getUsers = asyncWrapper(async (_req, res) => {
 export const getUser = asyncWrapper(async (req, res) => {
   const { user_id } = req.params;
 
-  const user = await userService.getUserById(user_id);
-
+  const user = await userService.getUserByProperty("_id", user_id);
   if (!user) throw createHttpError(404, "Requested user not found");
+
+  if (!userPermission.canGet(req.user, user))
+    throw createHttpError(401, "You don't have access to this resource");
 
   return res.status(200).json({ user: user.toObject() });
 });
 
 export const postUser = asyncWrapper(async (req, res) => {
-  const { email, password, username, roles } = req.body;
+  if (!userPermission.canCreate(req.user))
+    throw createHttpError(401, "You don't have access to create this resource");
+
+  const { email, password, username, role } = req.body;
 
   if (!email || !password || !username)
     throw createHttpError(400, "Please provide valid input");
@@ -28,7 +37,7 @@ export const postUser = asyncWrapper(async (req, res) => {
     email,
     password,
     username,
-    roles,
+    role,
   });
 
   return res
@@ -38,28 +47,41 @@ export const postUser = asyncWrapper(async (req, res) => {
 
 export const patchUser = asyncWrapper(async (req, res) => {
   const { user_id } = req.params;
-  const { email, username, roles } = req.body;
-  if (!email && !username && !roles)
-    throw createHttpError(400, "you have to update at least one property");
+  const { email, username, role } = req.body;
+
+  let user = await userService.getUserByProperty("_id", user_id);
+  if (!user) throw createHttpError(404, "Requested user not found");
+
+  if (!userPermission.canUpdate(req.user, user))
+    throw createHttpError(401, "You don't have access to update this resource");
+
+  if (!email && !username && !role)
+    throw createHttpError(400, "You have to update at least one property");
 
   const updates: Record<string, any> = {};
   if (email) updates.email = email;
   if (username) updates.username = username;
-  if (roles) updates.roles = roles;
+  if (role) updates.roles = role;
 
   // update user
-  const user = await userService.updateUserById(user_id, updates);
-
-  if (!user) throw createHttpError(404, "requested user not found");
+  user = await userService.updateUserById(user_id, updates);
 
   return res.status(200).json({ user: user.toObject() });
 });
 
 export const deleteUser = asyncWrapper(async (req, res) => {
   const { user_id } = req.params;
-  const user = await userService.deleteUserById(user_id);
 
+  let user = await userService.getUserByProperty("_id", user_id);
   if (!user) throw createHttpError(404, "requested user not found");
+
+  if (!userPermission.canDelete(req.user, user))
+    throw createHttpError(
+      401,
+      "You don't have any access to Delete this resource"
+    );
+
+  user = await userService.deleteUserById(user_id);
 
   return res.status(200).json({ userId: user_id });
 });
